prep for functional test
