+++ test/data/scaled-test-data.mtz
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:40dffda3f8eac31d37ecc8f2080ac4eb6f02c9616d2a084740ab9913ded86fc7
-size 2654416
+oid sha256:f0dc8d7378fd5a6cf93e8cfa761e44cc66c53733a48124acb05fc12a6c185439
+size 3649192
